Add files via upload
This class requires a CSV-file. Once you have one downloaded it makes a real-time prediction on the wished time horizon.
I do not recommend making trading decisions based on predictions from the algorithm and I am not responsible for how you use it or make trades from it. Have fun and feel free to implement new things or just copy the code for learning.